Add support for animated gif output.
As well as static images it is now possible to create animated gif
outputs, for example showing intermediate steps. Most of the heavy
lifting is done by GIFEncoder, but there are some helper methods
to aid setup and frame capture.

The API remains synchronous so this approach will not translate to
the browser.